requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
sentence-transformers
pypdf
ollama
chromadb
python-dotenv
PyYAML
pylance
